requirements.txt
amqp==5.2.0
appdirs==1.4.4
APScheduler==3.10.4
billiard==4.2.1
celery==5.5.3
coreapi==2.3.3
Django==4.2.25
django-appconf==1.0.6
django-codemirror2==0.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-filter==24.3
django-guardian==2.4.0
django-imagekit==5.0.0
django-redis==5.4.0
django-webpack-loader==0.7.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
djangorestframework-guardian==0.3.0
drf-nested-routers==0.94.1
gunicorn==23.0.0
geodeep==0.9.8
kombu==5.5.2
Markdown==3.7
olefile==0.47
openapi-codec==1.3.2
packaging==24.2
piexif==1.1.3
pilkit==3.0
Pillow==11.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pydantic==2.10.4
pyodm==1.5.11
pyparsing==3.2.1
pytz==2024.2
rcssmin==1.1.2
redis==5.2.1
requests-toolbelt==1.0.0
urllib3==1.26.20
requests>=2.32.3
rfc3987==1.3.8
rjsmin==1.2.3
simplejson==3.19.3
six==1.17.0
strict-rfc3339==0.7
tzlocal==5.2
uritemplate==4.1.1
vine==5.1.0
webcolors==24.11.1
rio-tiler==6.7.0
rio-color==2.0.1
rio-cogeo==5.3.3
rasterio==1.4.3
numpy==1.26.4
scipy==1.14.1
drf-yasg==1.21.8
zipstream-ng==1.8.0
